Extract CouponFieldArea into its standalone file.
