+++ client/my-sites/checkout/src/components/wp-checkout-order-review.tsx
@@ -3,11 +3,10 @@ import {
 	isDomainRegistration,
 	isDomainTransfer,
 } from '@automattic/calypso-products';
-import { FormStatus, useFormStatus } from '@automattic/composite-checkout';
 import { useShoppingCart } from '@automattic/shopping-cart';
 import { styled, joinClasses } from '@automattic/wpcom-checkout';
 import { useTranslate } from 'i18n-calypso';
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useCallback } from 'react';
 import isAkismetCheckout from 'calypso/lib/akismet/is-akismet-checkout';
 import { hasP2PlusPlan } from 'calypso/lib/cart-values/cart-items';
 import useCartKey from 'calypso/my-sites/checkout/use-cart-key';
@@ -19,7 +18,7 @@ import { currentUserHasFlag, getCurrentUser } from 'calypso/state/current-user/s
 import { getWpComDomainBySiteId } from 'calypso/state/sites/domains/selectors';
 import getSelectedSite from 'calypso/state/ui/selectors/get-selected-site';
 import { useCheckoutV2 } from '../hooks/use-checkout-v2';
-import Coupon from './coupon';
+import CouponFieldArea from './coupon-field-area';
 import { WPOrderReviewLineItems, WPOrderReviewSection } from './wp-order-review-line-items';
 import type { OnChangeItemVariant } from './item-variation-picker';
 import type { CouponFieldStateProps } from '../hooks/use-coupon-field-state';
@@ -28,7 +27,6 @@ import type {
 	ResponseCart,
 	RemoveProductFromCart,
 	ReplaceProductInCart,
-	CouponStatus,
 } from '@automattic/shopping-cart';
 
 const SiteSummary = styled.div`
@@ -43,29 +41,6 @@ const SiteSummary = styled.div`
 
 	@media ( ${ ( props ) => props.theme.breakpoints.tabletUp } ) {
 		margin-top: -8px;
-	}
-`;
-
-const CouponLinkWrapper = styled.div< { shouldUseCheckoutV2: boolean } >`
-	${ ( props ) => ( props.shouldUseCheckoutV2 ? `font-size: 12px;` : `font-size: 14px;` ) }
-`;
-
-const CouponAreaWrapper = styled.div< { shouldUseCheckoutV2: boolean } >`
-	padding-bottom: ${ ( props ) => ( props.shouldUseCheckoutV2 ? '12px' : 'inherit' ) };
-`;
-
-const CouponField = styled( Coupon )``;
-
-const CouponEnableButton = styled.button< { shouldUseCheckoutV2: boolean } >`
-	cursor: pointer;
-	text-decoration: underline;
-	color: ${ ( props ) => props.theme.colors.highlight };
-
-	&.wp-checkout-order-review__show-coupon-field-button {
-		${ ( props ) => ( props.shouldUseCheckoutV2 ? `font-size: 12px` : `font-size: 14px;` ) }
-	}
-	:hover {
-		text-decoration: none;
 	}
 `;
 
@@ -231,64 +206,6 @@ export default function WPCheckoutOrderReview( {
 				) }
 			</div>
 		</>
-	);
-}
-
-function CouponFieldArea( {
-	isCouponFieldVisible,
-	setCouponFieldVisible,
-	isPurchaseFree,
-	couponStatus,
-	couponFieldStateProps,
-}: {
-	isCouponFieldVisible: boolean;
-	setCouponFieldVisible: ( visible: boolean ) => void;
-	isPurchaseFree: boolean;
-	couponStatus: CouponStatus;
-	couponFieldStateProps: CouponFieldStateProps;
-} ) {
-	const { formStatus } = useFormStatus();
-	const translate = useTranslate();
-	const { setCouponFieldValue } = couponFieldStateProps;
-	const shouldUseCheckoutV2 = useCheckoutV2() === 'treatment';
-
-	useEffect( () => {
-		if ( couponStatus === 'applied' ) {
-			// Clear the field value when the coupon is applied
-			setCouponFieldValue( '' );
-		}
-	}, [ couponStatus, setCouponFieldValue ] );
-
-	if ( isPurchaseFree || couponStatus === 'applied' ) {
-		return null;
-	}
-
-	if ( isCouponFieldVisible ) {
-		return (
-			<CouponAreaWrapper shouldUseCheckoutV2={ shouldUseCheckoutV2 }>
-				<CouponField
-					id="order-review-coupon"
-					disabled={ formStatus !== FormStatus.READY }
-					couponStatus={ couponStatus }
-					couponFieldStateProps={ couponFieldStateProps }
-				/>
-			</CouponAreaWrapper>
-		);
-	}
-
-	return (
-		<CouponAreaWrapper shouldUseCheckoutV2={ shouldUseCheckoutV2 }>
-			<CouponLinkWrapper shouldUseCheckoutV2={ shouldUseCheckoutV2 }>
-				{ translate( 'Have a coupon? ' ) }{ ' ' }
-				<CouponEnableButton
-					className="wp-checkout-order-review__show-coupon-field-button"
-					onClick={ () => setCouponFieldVisible( true ) }
-					shouldUseCheckoutV2={ shouldUseCheckoutV2 }
-				>
-					{ translate( 'Add a coupon code' ) }
-				</CouponEnableButton>
-			</CouponLinkWrapper>
-		</CouponAreaWrapper>
 	);
 }
